backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
fyers-apiv3>=3.0.0
websockets>=12.0
